requirements.txt
urwid==2.0.1
pytest==3.4.2
boto3==1.6.11
mock==2.0.0
pytest==3.4.2
PyYAML==3.12
